+++ src/cv/cvemd.cpp
@@ -628,12 +628,12 @@ icvFindBasicVariables( float **cost, char **is_x,
             {
                 float cur_v_val = cur_v->val;
 
-                j = cur_v - v;
+                j = (int)(cur_v - v);
                 /* find the variables in column j */
                 prev_u = &u0_head;
                 for( cur_u = u0_head.next; cur_u != 0; )
                 {
-                    i = cur_u - u;
+                    i = (int)(cur_u - u);
                     if( is_x[i][j] )
                     {
                         /* compute u[i] */
@@ -665,14 +665,14 @@ icvFindBasicVariables( float **cost, char **is_x,
                 float *_cost;
                 char *_is_x;
 
-                i = cur_u - u;
+                i = (int)(cur_u - u);
                 _cost = cost[i];
                 _is_x = is_x[i];
                 /* find the variables in rows i */
                 prev_v = &v0_head;
                 for( cur_v = v0_head.next; cur_v != 0; )
                 {
-                    j = cur_v - v;
+                    j = (int)(cur_v - v);
                     if( _is_x[j] )
                     {
                         /* compute v[j] */
@@ -974,13 +974,13 @@ icvRussel( CvEMDState * state )
         prev_u = &u_head;
         for( cur_u = u_head.next; cur_u != 0; cur_u = cur_u->next )
         {
-            i = cur_u - u;
+            i = (int)(cur_u - u);
             float *delta_row = delta[i];
 
             prev_v = &v_head;
             for( cur_v = v_head.next; cur_v != 0; cur_v = cur_v->next )
             {
-                j = cur_v - v;
+                j = (int)(cur_v - v);
                 if( min_delta > delta_row[j] )
                 {
                     min_delta = delta_row[j];
@@ -1006,7 +1006,7 @@ icvRussel( CvEMDState * state )
         {
             for( cur_v = v_head.next; cur_v != 0; cur_v = cur_v->next )
             {
-                j = cur_v - v;
+                j = (int)(cur_v - v);
                 if( cur_v->val == cost[min_i][j] )      /* column j needs updating */
                 {
                     float max_val = -CV_EMD_INF;
@@ -1035,7 +1035,7 @@ icvRussel( CvEMDState * state )
         {
             for( cur_u = u_head.next; cur_u != 0; cur_u = cur_u->next )
             {
-                i = cur_u - u;
+                i = (int)(cur_u - u);
                 if( cur_u->val == cost[i][min_j] )      /* row i needs updating */
                 {
                     float max_val = -CV_EMD_INF;
