+++ src/highgui/grfmt_sunras.cpp
@@ -224,7 +224,7 @@ bool  GrFmtSunRasterReader::ReadData( uchar* data, int step, int color )
                
                 for(;;)
                 {
-                    int max_count = line_end - tsrc;
+                    int max_count = (int)(line_end - tsrc);
                     int code = 0, len = 0, len1 = 0;
 
                     do
@@ -292,7 +292,7 @@ bad_decoding_1bpp:
 
                 for(;;)
                 {
-                    int max_count = line_end - data;
+                    int max_count = (int)(line_end - data);
                     int code = 0, len = 0, len1;
                     uchar* tsrc = src;
 
@@ -308,7 +308,7 @@ bad_decoding_1bpp:
                     }
                     while( (max_count -= nch) > 0 );
 
-                    len1 = tsrc - src;
+                    len1 = (int)(tsrc - src);
 
                     if( len1 > 0 )
                     {
@@ -422,7 +422,7 @@ bool  GrFmtSunRasterWriter::WriteImage( const uchar* data, int step,
     
     if( m_strm.Open( m_filename ) )
     {
-        m_strm.PutBytes( fmtSignSunRas, strlen(fmtSignSunRas) );
+        m_strm.PutBytes( fmtSignSunRas, (int)strlen(fmtSignSunRas) );
         m_strm.PutDWord( width );
         m_strm.PutDWord( height );
         m_strm.PutDWord( channels*8 );
